% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql_analyst_server.R
\name{read_sql_analyst_Server}
\alias{read_sql_analyst_Server}
\title{Read Data
from Midlands and Lancashire CSU
Business Intelligence Analyst SQL Server}
\usage{
read_sql_analyst_Server(sql_query)
}
\arguments{
\item{sql_query}{A SQL Query to read data from the
Midlands and Lancashire CSU Analyst Server.}
}
\value{
A dataset taken from Midlands and Lancashire CSU SQL Analyst Server
}
\description{
This function allows the user to connect
to the commonly used analyst server directly to access data
}
\details{
The only input needed is the SQL query used to generate the data
}
\examples{



library(MLCSUr)

# write a SQL query to import the data you require
example_sql_query <- "
                    SELECT TOP (1000) *
                    FROM EAT_Reporting.dbo.tbInpatientEpisodes
                   "

# pass the SQL query to the read_sql_analyst_Server
#     function to read the data in
episodesData <- read_sql_analyst_Server(example_sql_query)

}
