% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSQLAnalystServer.R
\name{readSQLAnalystServer}
\alias{readSQLAnalystServer}
\title{Read data from Midlands and Lancashire CSU Analyst SQL Server}
\usage{
readSQLAnalystServer(SQLQuery)
}
\arguments{
\item{SQLQuery}{A SQL Query to read data from the Midlands and Lancashire CSU Analyst Server.}
}
\value{
A dataset taken from Midlands and Lancashire CSU SQL Analyst Server
}
\description{
Read Data from Midlands and Lancashire CSU Business Intelligence Analyst SQL Server
}
\details{
This function allows the user to connect to the commonly used analyst server directly to access data

The only input needed is the SQL query used to generate the data
}
\examples{

library(MLCSUr)

# write a SQL query to import the data you require
exampleSQLQuery <- "SELECT TOP (1000) * FROM EAT_Reporting.dbo.tbInpatientEpisodes"

# pass the SQL query to the readSQLAnalystServer function to read the data in
episodesData <- readSQLAnalystServer(exampleSQLQuery)

}
